Reorganized io functions and splitted them into pharmacophore writer, reader and mol files
